chore(storybook): Move non-component stories out of design-system dir
