+++ .agents/skills/storybook/SKILL.md
@@ -6,5 +6,5 @@ description: Use when writing or reviewing Storybook stories (`.stories.tsx`) fo
 # Storybook Component Stories
 
 Before writing or reviewing Storybook stories, read and follow the canonical
-[Writing Good Stories](../../../web/storybook/docs/WritingGoodStories.mdx)
+[Writing Good Stories](../../../web/.storybook/docs/WritingGoodStories.mdx)
 guide.
