+++ web/.storybook/docs/ThemeTokens/Typography.stories.tsx
@@ -1,4 +1,4 @@
-import preview from "../../../../.storybook/preview";
+import preview from "../../preview";
 import { Typography as TypographyPage } from "./Typography";
 
 const meta = preview.meta({
