+++ web/.storybook/storybook.css
@@ -1,0 +1,3 @@
+@import "../src/styles/globals.css";
+
+@source "./docs";
