+++ web/.structure-baseline.json
@@ -8,9 +8,6 @@
    "src/components/date-picker.tsx: filename not PascalCase; 3 components: DatePicker, DatePickerWithRange, TimeRangePicker",
    "src/components/date-range-dropdowns.tsx: filename not PascalCase",
    "src/components/deleteButton.tsx: filename not PascalCase; 7 components: DeleteButton, DeleteTraceButton, DeleteDatasetButton, DeleteDashboardButton, DeleteMonitorButton, DeleteEvalConfigButton, DeleteEvaluationModelButton",
-   "src/components/design-system/ThemeTokens/docsChrome.tsx: filename not PascalCase; 5 components: Eyebrow, SpecChip, DocsPageHeader, DocsSection, CropFrame",
-   "src/components/design-system/ThemeTokens/Layout.tsx: 7 components: RadiiSection, LayoutOffsetsSection, LayeringSection, LayerSystemSection, AnimationsSection, BreakpointsNote, Layout",
-   "src/components/design-system/ThemeTokens/shared.tsx: filename not PascalCase; 5 components: Swatch, EmptyCell, TokenRow, TokenSection, CollapsedSection",
    "src/components/error-page.tsx: filename not PascalCase; 2 components: ErrorPage, ErrorPageWithSentry",
    "src/components/grouped-score-badge.tsx: filename not PascalCase",
    "src/components/layouts/app-layout/index.tsx: filename not PascalCase",
@@ -202,7 +199,6 @@
    "src/app/layout.tsx: extra exports: metadata",
    "src/components/ChatMessages/messageSearch/context.tsx: extra exports: useMessageSearchActions, useOptionalMessageSearchActions, useOptionalMessageSearchPageId, useMessageSearch, useSyncMessageSearchMessages",
    "src/components/date-picker.tsx: extra exports: nextRangeForDayClick, isRangeWithinMaxDuration",
-   "src/components/design-system/ThemeTokens/shared.tsx: extra exports: parsed, rootEntries, buildContext, useIsDarkTheme, useTokenContexts, PAGE_TOKEN_MATCHERS, allTokenNames, pageForToken, unassignedTokens, findTokenDeclaration",
    "src/components/editor/CodeMirrorEditor.tsx: extra exports: getPromptVariableDiagnostics, applyCodeMirrorSearchQuery, setActiveSearchMarkCodeMirrorRange, unsetActiveSearchMarkCodeMirrorRange",
    "src/components/grouped-score-badge.tsx: extra exports: groupScoresByName",
    "src/components/ItemBadge.tsx: extra exports: renderFilterIcon",
