--- web/storybook/docs/Overview.mdx
@@ -1,27 +1,0 @@
-import { Meta } from "@storybook/addon-docs/blocks";
-
-<Meta />
-
-# Overview
-
-Storybook is the shared workspace for documenting Langfuse's interface and
-developing UI components in isolation. It keeps design guidance and component
-examples easy to discover without requiring the full application.
-
-Before modifying content in this Storybook, please read the [Writing Good Stories](?path=/docs/design-writing-good-stories--docs) page.
-
-## Structure
-
-Our Storybook is split into two sections with different purposes:
-
-#### Design
-
-Design contains the curated documentation for Langfuse's visual language,
-patterns, and conventions. Use it for guidance that should remain stable and
-help contributors make consistent product decisions.
-
-#### Playground
-
-Playground contains isolated component stories and supporting technical
-documentation. Use it to develop, test, and explore components independently
-from the application. New component stories belong here by default.
